+++ test/fixtures.js
@@ -1,7 +1,0 @@
-module.exports = {
-    playerOne: {
-        email: 'FrederickKBonner@dayrep.com',
-        name: 'Frederick K. Bonner',
-        id: 'PL-fdiTUIO'
-    }
-};