+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 [0.1.0]
 -------
-- [ ] Add `factory.create` subs
+- [x] Add initial commands loading
+- [x] Add `region.create` subs
+- [x] Add `player.register` subs
+- [x] Add `factory.create` subs
 - [ ] Add `produce` subs